Reduces the diff on test_objectives.py to zero and keeps DreamerV3 tests self-contained alongside the other dedicated test files.
